requirements.txt
# FSKU LLM Challenge - 통일된 의존성 관리
# 날짜: August 10, 2025
# 기준: 대회 환경 (Ubuntu 22.04, Python 3.10, PyTorch 2.1.0, CUDA 11.8)

# -- Core ML & Frameworks (대회 지정 및 필수 라이브러리) --
torch==2.1.0
transformers==4.41.2
accelerate==0.30.1
peft==0.11.1
trl==0.9.6
vllm==0.5.4

# -- RAG: Embedding & Vector DB --
sentence-transformers==2.7.0
faiss-cpu==1.8.0  # GPU 메모리 확보를 위해 CPU 버전 사용
bm25s==0.2.2

# -- Model Optimization --
auto-gptq==0.7.1
bitsandbytes==0.43.1

# -- Data Processing & Utilities --
pandas==2.2.2
langchain==0.2.1
PyMuPDF==1.24.1
beautifulsoup4==4.12.3
lxml==5.2.2
ftfy==6.2.0
chardet==5.2.0
konlpy==0.6.0

# -- Monitoring & Testing --
tqdm==4.66.4
wandb==0.17.0
pytest==8.2.0

# -- Development Tools --
black==23.7.0
ruff==0.0.285
pre-commit==3.3.3